Fixed makefile and added ast.c and base.c to compile list
Update to fix compile issue

Added visitors and parsing for functions/procs.

+++ symbol_table.c
@@ -109,13 +109,13 @@ symbol_print(Symbol *symbol)
         return;
     }
 
-    printf("Symbol: %x\n", symbol);
+    printf("Symbol: %x\n", (unsigned)symbol);
     printf("name: %s\n", symbol->name);
     printf("type: %d\n", symbol->type);
     printf("value:");
     value_print(stdout, &symbol->value, symbol->type);
     printf("\ndeclaration line: %d\n", symbol->decl_linenum);
-    printf("next: %x\n\n", symbol->next);
+    printf("next: %x\n\n", (unsigned)symbol->next);
 }
 
 void
